gateway/requirements.txt
Flask
grpcio-tools
prometheus-client